services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
databases[asyncpg]==0.9.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
httpx==0.27.2
openai==1.54.0
anthropic==0.40.0
json-repair==0.30.3
pyyaml>=6.0
